# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Notes API",
    "version": "1.0.0",
    "description": "API documentation for Smart Notes application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "Note": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Note ID",
            "example": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp",
            "example": "2024-01-01T12:00:00Z"
          },
          "notebook_id": {
            "type": "integer",
            "description": "ID of the notebook this note belongs to",
            "example": 1
          },
          "content": {
            "type": "string",
            "nullable": true,
            "description": "Note content",
            "example": "This is my first note"
          },
          "order_index": {
            "type": "integer",
            "nullable": true,
            "description": "Order position of the note",
            "example": 0
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last update timestamp",
            "example": "2024-01-01T12:00:00Z"
          }
        },
        "required": [
          "id",
          "created_at",
          "notebook_id"
        ]
      },
...
      "Notebook": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Notebook ID",
            "example": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp",
            "example": "2024-01-01T12:00:00Z"
          },
          "title": {
            "type": "string",
            "description": "Notebook title",
            "example": "My First Notebook"
          },
          "description": {
            "type": "string",
            "description": "Notebook description",
            "example": "This is a notebook for personal notes"
          },
          "color": {
            "type": "string",
            "description": "Notebook color (hex code or color name)",
            "example": "#3B82F6"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp",
            "example": "2024-01-01T12:00:00Z"
          },
          "folder_id": {
            "type": "integer",
            "description": "ID of the folder this notebook belongs to",
            "example": 1
          },
          "order_index": {
            "type": "integer",
            "description": "Order position of the notebook",
            "example": 0
          }
        },
        "required": [
          "id",
          "created_at",
          "title",
          "description",
          "color",
          "updated_at",
          "folder_id",
          "order_index"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/notebooks": {
      "get": {
        "summary": "Get all notebooks (optionally filtered by folder)",
        "tags": [
          "Notebooks"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "folder_id",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Filter notebooks by folder ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of notebooks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notebooks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notebook"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "description": "Total number of notebooks"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new notebook",
        "tags": [
          "Notebooks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "color",
                  "folder_id",
                  "order_index"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Notebook title",
                    "example": "My First Notebook"
                  },
                  "description": {
                    "type": "string",
                    "description": "Notebook description",
                    "example": "This is a notebook for personal notes"
                  },
                  "color": {
                    "type": "string",
                    "description": "Notebook color (hex code or color name)",
                    "example": "#3B82F6"
                  },
                  "folder_id": {
                    "type": "integer",
                    "description": "ID of the folder this notebook belongs to",
                    "example": 1
                  },
                  "order_index": {
                    "type": "integer",
                    "description": "Order position of the notebook",
                    "example": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notebook created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Notebook created successfully"
                    },
                    "notebook": {
                      "$ref": "#/components/schemas/Notebook"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/notebooks/count/{id}": {
      "get": {
        "summary": "Get count of all notebooks filtered by folder",
        "tags": [
          "Notebooks"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "folder_id",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Filter notebooks by folder ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Count of notebooks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer",
                      "description": "Total number of notebooks"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/notebooks/{id}": {
      "get": {
        "summary": "Get a notebook by ID",
        "tags": [
          "Notebooks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The notebook ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notebook details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notebook"
                }
              }
            }
          },
          "404": {
            "description": "Notebook not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a notebook",
        "tags": [
          "Notebooks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The notebook ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Updated notebook title",
                    "example": "Updated Notebook"
                  },
                  "description": {
                    "type": "string",
                    "description": "Updated notebook description",
                    "example": "Updated description"
                  },
                  "color": {
                    "type": "string",
                    "description": "Updated notebook color",
                    "example": "#10B981"
                  },
                  "folder_id": {
                    "type": "integer",
                    "description": "Updated folder ID",
                    "example": 2
                  },
                  "order_index": {
                    "type": "integer",
                    "description": "Updated order position",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notebook updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Notebook updated successfully"
                    },
                    "notebook": {
                      "$ref": "#/components/schemas/Notebook"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a notebook",
        "tags": [
          "Notebooks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The notebook ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notebook deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Notebook deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/notes": {
      "get": {
        "summary": "Get all notes (optionally filtered by notebook)",
        "tags": [
          "Notes"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "notebook_id",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Filter notes by notebook ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Note"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "description": "Total number of notes"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new note",
        "tags": [
          "Notes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "notebook_id"
                ],
                "properties": {
                  "notebook_id": {
                    "type": "integer",
                    "description": "ID of the notebook this note belongs to",
                    "example": 1
                  },
                  "content": {
                    "type": "string",
                    "description": "Note content",
                    "example": "This is my first note"
                  },
                  "order_index": {
                    "type": "integer",
                    "description": "Order position of the note",
                    "example": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Note created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Note created successfully"
                    },
                    "note": {
                      "$ref": "#/components/schemas/Note"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/notes/{id}": {
      "get": {
        "summary": "Get a note by ID",
        "tags": [
          "Notes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The note ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Note details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a note",
        "tags": [
          "Notes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The note ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "Updated note content",
                    "example": "Updated note content"
                  },
                  "notebook_id": {
                    "type": "integer",
                    "description": "Updated notebook ID",
                    "example": 2
                  },
                  "order_index": {
                    "type": "integer",
                    "description": "Updated order position",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Note updated successfully"
                    },
                    "note": {
                      "$ref": "#/components/schemas/Note"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a note",
        "tags": [
          "Notes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The note ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Note deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Note deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}